feat(client): Separated hype creation process
