+++ services/client/src/pages/AddHypePool/DetailsForm/index.ts
@@ -1,0 +1,2 @@
+export * from './DetailsForm';
+export * from './DetailsForm.effects';
