+++ services/client/src/pages/AddHypePool/RewardForm/index.ts
@@ -1,0 +1,2 @@
+export * from './RewardForm';
+export * from './RewardForm.effects';
