+++ services/client/tsconfig.json
@@ -26,6 +26,6 @@
     "isolatedModules": true,
     "jsx": "react-jsx"
   },
-  "include": ["src/**/*", "__tests__/**/*"],
+  "include": ["src/**/**/*", "__tests__/**/*"],
   "exclude": ["node_modules", "build", "*.d.ts"]
 }
